refactor: renamed database file for clarity
